benchmarks/WiseEdit/requirements.txt
# core dependencies
openai==2.9.0
pillow==12.0.0
pandas==2.3.3
tqdm==4.67.1